드가자

+++ chapter04_MVC/src/main/java/org/jae/service/ReplyServiceImpl.java
@@ -3,9 +3,11 @@ package org.jae.service;
 import java.util.List;
 
 import org.jae.domain.ReplyVO;
+import org.jae.mapper.BoardMapper;
 import org.jae.mapper.ReplyMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.extern.log4j.Log4j;
 
@@ -15,9 +17,16 @@ public class ReplyServiceImpl implements ReplyService {
 	@Autowired
 	private ReplyMapper mapper;
 	
+	@Autowired BoardMapper boardMapper;
+	
+
+	
+	@Transactional
 	@Override
 	public int register(ReplyVO rvo) {
 	   log.info("register..." + rvo);
+	   // replycnt 값 증가
+	    boardMapper.updateReplyCnt(1, rvo.getBno());
 		return mapper.insert(rvo);
 	}
 	@Override
@@ -35,9 +44,25 @@ public class ReplyServiceImpl implements ReplyService {
 		log.info("modify...");
 		return mapper.update(rvo);
 	}
+
+	@Transactional
 	@Override
 	public int remove(int rno) {
-		log.info("remove.......");
-		return mapper.delete(rno);
+	    
+	    log.info("remove.......");
+
+	  
+	    ReplyVO rvo = mapper.read(rno);
+	    
+	    if (rvo == null) {
+	        return 0;
+	    }
+
+	    // 게시글의 댓글 수 감소
+	    boardMapper.updateReplyCnt(-1, rvo.getBno());
+
+	    // 댓글 삭제
+	    return mapper.delete(rno);
 	}
+
 }
